requirements.txt
fastapi
starlette
pytest
httpx
PyJWT
uvicorn[standard]
